Implement passive mode. Taken from the updated manpage:
"Enable passive mode, where the server never replies requests, but acts upon
them. It's useful when you want to have a live database mirror, in case the
main node fails. It still requires a manual restart to became active, but at
least you have a recent up-to-date database. Be aware that it's not widely
tested (although it should work fine), and do not start an active and a
passive server in the same machine (it misbehaves under some circumnstances,
and doesn't make much sense anyway)."