Use conditional waits for waking up the db thread. This patch changes the db thread wakeup mechanism to be conditional waits, instead of just nanosleep as it used to be.
It has several advantages, the main one being the reduction of the latency for
all synchronous operations, including get which on cache misses is always
synchronous, and reduced CPU time on idle.

The wakeup is done in a conditional way, to avoid waking the thread up when an
asynchronous operation is queued. As there is no need for an inmediate
response, we skip the wakeup, because it will notice it when the timer
expired, reducing the response latency.

To verify this conditional wakeup was worthy, I ran "test2d 10000 10 10" and
compared the set and del times:

	cond	always
	wakeup	wakeup
---------------------------
set	2083847 2804973
del	1986236 2605961

It's a clear win for little effort.