Mark as internal some variables in the Python bindings.
The Python bindings exposes "get", "set", "del" functions, which are
actually direct calls to the low level module and are not meant to
be used (they are undocumented, the return values are strange,
etc.).

This patch makes them begin with an underscore to note the fact they
are internal and not made to be used.