Allow CFLAGS override in the Makefiles.
This patch allows the CFLAGS variable to be overriden by the user when
calling make, for instance "make CFLAGS=-Os".

Previously, that would have broken the build because the needed CFLAGS
were not included (-D_XOPEN_SOURCE=500, for instance, is mandatory and
the user would have to include that in its CFLAGS).

Signed-off-by: Alberto Bertogli <albertito@gmail.com>