Reduce tipc_recv()'s buffer size.
There is a small but noticeable performance gain if we reduce the
tipc_recv()'s buffer size from 128k to 68k. Since the maximum TIPC packet
size is 66000 bytes, this should be enough for even the biggest packets.

Making the buffer static and outside the function was tested but it
performed worse than the current stack-based one.

It's still ugly, but at least it uses almost half the stack size.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>