Implement server-side TCP support.
This adds basic TCP support to the nmdb server.
It's still missing features (parametrization), performance optimization
and cleanups, but it's fully functional, it runs the tests, handles
invalid inputs and makes valgrind happy.

The main problem with it is that it can't handle having more than one
incoming message on the same socket yet, but it will come on the next
patches.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>