Isolate message parsing from network message receiving.
This patch cleanly splits message parsing from message receiving.
While on TIPC the message received IS the one to parse, on TCP we have
a thin but troublesome message wrapper that we don't want to mix with
our real code.

It also simplifies the code a lot, and makes it easier to add new
protocols in the future.


Signed-off-by: Alberto Bertogli <albertito@gmail.com>