Allow to compile-out TIPC and TCP in the server.
To compile either TCP or TIPC out, use a "make" variable from the command
line, like "make ENABLE_TCP=0" or "make ENABLE_TIPC=0". You can disable
both to get a very useful application.

In the future we might want to revisit the selection method, but this one
is simple and good enough for now.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>