Improve libnmdb's TCP support.
Until now, we did two send() and two recv(), one for the message length,
and other for the message itself. This slowed TCP down a lot.

This patch makes recv() handle the common case where we get the whole
thing in one go, and adds generic send() infraestructure to handle
transport protocols which need to get room for a header before sending
the packet.

This improved TCP performance in a significant way, making it 1/3 faster
than before, although it's still slower than TIPC on localhost.

Tested with all testcases and valgrind.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>