Rewrite test building.
The tests were getting too similar and boring to write.
This patch simplifies the situation by having only two tests with generic
code, and then specifying the protocol and access mode on build-time.

It also changes the Makefile in favour of a building script so we can
build all the combinations easily.

This gets us 18 tests, from the combination of all the following:
 - TCP/TIPC/Multiple (both at the same time).
 - Normal/Cache/Sync access.
 - Test code 1 and 2.


Signed-off-by: Alberto Bertogli <albertito@gmail.com>