Use static allocation in TCP's receive common path.
To avoid too many expensive malloc()s, and also to make the code
simpler (with regards to the current status), use a static buffer
to use in TCP's common receive path.

This shows a slight increase in performance, but it's worth it just for
the simplified code, even though the common buffer is not the nicest
thing on earth.

I hope I can get rid of it with an optimized allocator, but that can wait.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>