Make nmdb_*get() return -1 on miss, instead of 0. Breaks the API.
Without this patch, it's impossible to tell between a miss and a key
with a value of "" (the empty string).

This fixes that by returning -1 (instead of 0) when there's a miss, and -2
(instead of -1) when there's an error.

It updates all bindings and tests.

Too bad it missed the 0.20 release.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>