Try harder to send() when we get EAGAIN.
With big packets, the receiver window might get exhausted and send()
would block, but as the fd is set in non-blocking mode, it returns
EAGAIN. This patch retries when send() fails in this way.

The proper way to fix this would be to add an event so we get notified
when the fd is available for writing, and retry the send; but this is
complex so leave it for when it's really needed.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>