Remove the 'STRICT' build option and add -pedantic as default.
The -ansi was wrong, because it set the standard to C90, overriding the
earlier -std=c99 option.

After we remove it, it makes no sense leaving -pedantic as an option; it
should be in the default CFLAGS, as we should be sticking to C99 and not
rely on any GNU extensions.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>