Improve and fix support for compiling out TIPC, UDP and TCP from libnmdb.
The support for compiling out TIPC, UDP and TCP in libnmdb was broken,
because while we did not include the code, we still had a dependency on
the header, because nmdb.h included the tipc and inet headers regardless
of the configuration options.

This patch fixes this by generating nmdb.h on build time, and setting
some internal constants accordingly. Not the nicest way on earth, but it's
not too dirty either, and it works.

It also fixes some issues with having UDP in and TCP out or viceversa,
because the server comparison function relied on TCP (it worked on UDP
just by chance because accesing the data through the TCP union member
is ABI compatible). Now there is only one union member for either, named
'in', accordingly to the headers and the sockaddr variant used.


Signed-off-by: Alberto Bertogli <albertito@gmail.com>