Delete all events properly on shutdown.
This fixes a bug reported by Arnaud Connois <arnaud@connois.com> that
caused a libevent assertion to raise on event_del().

That was because we were deleting tipc_evt, but it may not be initialized
properly if ENABLE_TIPC=0.

It triggered on Fedora Core 5 installs but not on Fedora 7 or Gentoo,
probably because of some differences in the build environment.

Thanks to Arnaud Connois for all the assistance tracking this down.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>