Use request flags instead of having lots of different requests
Currently, on the wire we have different requests for what actually are
alternatives of the same request (for instance, a set request can be
either REQ_SET_SYNC, REQ_SET_ASYNC or REQ_CACHE_SET). Internally, they're
handled with the same code, except the alternatives are considered.

This patch removes these different requests and adds flags instead, both
on the protocol and on the internal API, which make the code more clear.

Because we're not in a stable release, upgrading the protocol version is
not worth the trouble.

Signed-off-by: Alberto Bertogli <albertito@gmail.com>