Move htonll() and ntohll() to a common, shared file
As we expect to use these functions in different places in the future, it
makes no sense anymore to have this code duplicated inside their users. So
move them to a common file, and use symlinks because it's not worth the
trouble of having a common directory for just to files (yet).

Signed-off-by: Alberto Bertogli <albertito@gmail.com>