python: Fix key selection when autopickling
When autopickling is enabled, we select the key using str(hash(key)),
which is stupid (and non portable) because the hash() of some classes is
just the object's memory address, which is obviously not stable across
invocations.

The fix is to just use pickle to select the key.

Thanks to Kelvin (lbh1976@gmail.com) for the bug report.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>