tests/python*: Fix random operation count calculation
The operation count is half of the number of keys, but in the cache-only
tests, the number of available keys to operate on can be less than that due to
cache pressure.

This can be reproduced by running the server with a small cache (e.g. -c 1)
and the tests with a large number of operations (e.g. 5000).

This patch fixes it by checking that we still have keys in the cache to
operate on.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>