nmdb: Put cache entries statically into the cache chain they belong to
Instead of handling the cache entries with malloc() and free() all the time,
and having the chain length as a member of the cache structure, make the
length a constant (which already was) and place a statically-allocated array
of entries inside each chain.

That reduces the costs of allocation, and makes it more cache-friendly. It has
a small but noticeable impact on performance.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>