nmdb: Change the cache hash function to Austin Appleby's MurmurHash2
After testing several hash functions, the one used in the cache was
changed to Austin Appleby's MurmurHash2.

MurmurHash2 is much faster than the one we use (10-25% in our benchmarks),
has nice and clean code, and is in the public domain.

We keep the old Jenkins' one-at-a-time function, and also add FNV and
Paul Hsieh's SuperFastHash, so the tests can be redone in the future if
needed.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>