Change nmdb's license to BOLA (Public domain) and update the license version
This commits re-licenses nmdb under the BOLA license (public domain). Since
I'm the sole copyright owner at the moment, it is OK to do the change.

Now the whole project is licensed under the same terms.

It also updates the other sub-projects to a more recent version of the
license.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>