Support building with multiple backends
This patch makes nmdb build with support with several backends, and lets the
user select which one to use at runtime.

This includes major changes to the internal database API, but luckily it never
goes outside the dbloop.c, and most changes are simple.

The Makefile has been extended to auto-detect available backends at build
time, so the user no longer needs to select which ones to use (although it is
still possible).

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>