libnmdb: Add Doxygen documentation to the public API
This patch adds Doxygen documentation to the library's public API. To do that,
it moves some structures and defines that were internal to the internal header
(which now needs to be generated at build time, so it was renamed to
internal.h.in).

Now the public header contains only the public API, and does not need to be
generated automatically anymore.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>