Add coverage tests
This patch adds a script that uses the other tests in an attempt to maximize
code coverage.

It is mostly useful to check which code is not covered by the other tests,
although the output is saved for manual inspection.

Scripts for generating nice reports with lcov are also included.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>