nmdb: Cast to avoid known const violation warnings in tdb and bdb backends
tdb and bdb APIs make it difficult to maintain const warranties, although
they do not modify the buffers.

This causes the compiler to emit a warning, which is distracting since it
is a known but harmless issue.

This patch adds casts to avoid the warnings. It was tested with gcc 4.4.3
and clang 1.1 (svn rev 98154).

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>