Add methods to iterate through all the keys in a server
This patch adds two functions, nmdb_firstkey() and nmdb_nextkey(), that
can be used to iterate through all the keys in a server.

It has some limitations (one server only, concurrent sets can cause
trouble, etc.), but it is needed for the replication utilities.

Only Python 2 bindings are included, the rest, along with the documentation,
will come in future patches.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>