nmdb: Add a signal handler to change to read-only mode
This patch adds a new signal handler that can be used to change to
read-only mode.

As SIGUSR1 is a nice signal to assign to this, but it is taken by the
log reopening. Since SIGHUP is the one traditionally used for that
purpose, we reassign them so now we have:

 - SIGHUP: reopen log file
 - SIGUSR1: switch to read-only mode
 - SIGUSR2: toggle passive mode

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>