nmdb: Improve check for broken values in cache_incr()
The code did not check that we do not give strtoll() a NULL value. We know it
should never happen, but it's better to check.

This patch modifies the previous test for 0-terminated strings (which actually
did check for val != NULL before accessing it), making it also return in the
case where val == NULL.

Caught by the clang-analyzer tool.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>