libnmdb: Remove redundant _ENABLE_X constants
There is no need to pass -DENABLE_X when building, the constants are readily
available in internal.h, but as _ENABLE_X.

This patch removes the redundancy by turning the internal.h constants into
ENABLE_X form, and not passing the -D when building. The code then gets the
constants from internal.h.

Thanks to Jay Sprenkle <jsprenkle@gmail.com> for the report.

Signed-off-by: Alberto Bertogli <albertito@blitiri.com.ar>